refactor(core): extract react.ts from index.ts (#663)
