+++ src/vanilla.ts
@@ -52,7 +52,7 @@ export type StateCreator<
   CustomStoreApi extends StoreApi<T> = StoreApi<T>
 > = (set: CustomSetState, get: CustomGetState, api: CustomStoreApi) => T
 
-function create<
+function createStore<
   TState extends State,
   CustomSetState,
   CustomGetState,
@@ -66,11 +66,11 @@ function create<
   >
 ): CustomStoreApi
 
-function create<TState extends State>(
+function createStore<TState extends State>(
   createState: StateCreator<TState, SetState<TState>, GetState<TState>, any>
 ): StoreApi<TState>
 
-function create<
+function createStore<
   TState extends State,
   CustomSetState,
   CustomGetState,
@@ -150,4 +150,4 @@ function create<
   return api as unknown as CustomStoreApi
 }
 
-export default create
+export default createStore
